LayoutTests/crypto/webkitSubtle/argument-conversion.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test crypto.webkitSubtle argument conversion");

jsTestIsAsync = true;

Promise.resolve(null).then(function() {
    debug("\nPassing algorithm name as a string object...");
    return crypto.webkitSubtle.digest(new String('sha-1'), new Uint8Array([]));
}).then(function(result) {
    testPassed("...succeeded");
    debug("\nPassing algorithm name as a string object in a dictionary...");
    return crypto.webkitSubtle.digest({name: new String('sha-1')}, new Uint8Array([]));
}).then(function(result) {
    testPassed("...succeeded");
    debug("\nPassing algorithm name as an object with toString");
    return crypto.webkitSubtle.digest({name: { toString:function() { return 'sha-1' } } }, new Uint8Array([]));
}).then(function(result) {
    testPassed("...succeeded");

    debug("\nPassing invalid data to digest()");
    shouldThrow("crypto.webkitSubtle.digest({name: 'sha-1'})");
    shouldThrow("crypto.webkitSubtle.digest({name: 'sha-1'}, null)");
    shouldThrow("crypto.webkitSubtle.digest({name: 'sha-1'}, 10)");
    shouldThrow("crypto.webkitSubtle.digest({name: 'sha-1'}, [10])");
    shouldThrow("crypto.webkitSubtle.digest({name: 'sha-1'}, [new Uint8Array([0])])");

    debug("\nPassing invalid algorithmIdentifiers to digest()");
    data = new Uint8Array([0]);
    shouldThrow("crypto.webkitSubtle.digest({ toString:function() { return 'sha-1' } }, data)"); // Algorithm normalization doesn't attempt to call toString.
    shouldThrow("crypto.webkitSubtle.digest({name: ''}, data)");
    shouldThrow("crypto.webkitSubtle.digest({name: null}, data)");
    shouldThrow("crypto.webkitSubtle.digest({name: undefined}, data)");
    shouldThrow("crypto.webkitSubtle.digest({name: 'sha'}, data)");
    shouldThrow("crypto.webkitSubtle.digest({name: 1}, data)");
    shouldThrow("crypto.webkitSubtle.digest('', data)");
    shouldThrow("crypto.webkitSubtle.digest(null, data)");
    shouldThrow("crypto.webkitSubtle.digest(undefined, data)");
    shouldThrow("crypto.webkitSubtle.digest(1, data)");
    shouldThrow("crypto.webkitSubtle.digest({}, data)");
}).then(finishJSTest);

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
